Fix test hang on Linux by providing required arguments to sleep process
The KillAllProcessesAsync_ShouldTerminateAllProcesses test was starting
/bin/sleep processes without the required duration argument on Linux,
causing the processes to be in an invalid state that couldn't be properly
terminated. This resulted in a 25-second hang followed by MSBuild crash.

Changed line 112-113 to pass GetLongRunningArguments() to StartProcess(),
ensuring sleep receives the required "60" duration argument on Linux.

Fixes GitHub Actions workflow run #19436268788 test timeout.